modificacion inicializacion de campos

+++ src/main/java/mx/axxib/aforedigitalgt/eml/RegisSinSalarioOut.java
@@ -22,7 +22,7 @@ import lombok.NoArgsConstructor;
 					columns = {
 						@ColumnResult(name = "num_solicitud", type = Integer.class),
 						@ColumnResult(name = "nss", type = String.class),
-						@ColumnResult(name = "sbc_tipo_a", type = Integer.class),
+						@ColumnResult(name = "sbc_tipo_a", type = Double.class),
 						@ColumnResult(name = "fecha_carga", type = Date.class),
 						@ColumnResult(name = "curp", type = String.class),
 						@ColumnResult(name = "primer_apellido", type = String.class),
@@ -35,7 +35,7 @@ public class RegisSinSalarioOut {
 
 	private Integer onNumSolicitud;
     private String onNss;
-    private Integer onSbcTipoA;
+    private Double onSbcTipoA;
     private Date odFechaCarga;
     private String ocCurp;
     private String ocPrimerApellido;
